Merge --mirror and many other features

+++ python/modules/Document.py
@@ -2,11 +2,21 @@
 
 from __future__ import annotations
 
-import json, re, os
-import Utils, Render, Alert, Html
-from typing import Tuple, Type, Callable
+import re, os, itertools
+import Utils, Render, Alert, Html, Filter
+from typing import Tuple, Type, Callable, Iterable
 import pyratemp, markdown
 from markdown_newtab_remote import NewTabRemoteExtension
+
+
+def WordCount(text: str) -> int:
+    "Return the approximate number of words in text"
+    words = re.split(r"\s+",text)
+    if len(words) > 1:
+        return len(words) - (not words[0]) - (not words[-1])
+            # Check if the first and last word are empty. Note: bool True = 1
+    else:
+        return len(words) - (not words[0])
 
 def RenderDocumentationFiles(aboutDir: str,destDir:str = "",pathToPrototype:str = "../",pathToBase = "../../",html:bool = True) -> list[Html.PageDesc]:
     """Read and render the documentation files. Return a list of PageDesc objects.
@@ -16,6 +26,7 @@ def RenderDocumentationFiles(aboutDir: str,destDir:str = "",pathToPrototype:str 
     pathToBase: the path to the base directory
     html: Render the file into html? - Leave in .md format if false.
     """
+    global gDocumentationWordCount
 
     aboutDir = Utils.PosixJoin(gOptions.documentationDir,aboutDir)
     if not destDir:
@@ -23,14 +34,18 @@ def RenderDocumentationFiles(aboutDir: str,destDir:str = "",pathToPrototype:str 
     sourceDir = aboutDir + "Sources"
     
     fileContents = {}
+    fileModified = {}
     for fileName in sorted(os.listdir(sourceDir)):
         sourcePath = Utils.PosixJoin(sourceDir,fileName)
 
         if not os.path.isfile(sourcePath) or not fileName.endswith(".md"):
             continue
 
+        fileModified[fileName] = Utils.ModificationDate(sourcePath)
         with open(sourcePath,encoding='utf8') as file:
-            fileContents[fileName] = file.read()
+            template = pyratemp.Template(file.read())
+            fileContents[fileName] = template(gOptions = gOptions,gDatabase = gDatabase)
+            gDocumentationWordCount += WordCount(fileContents[fileName])
             
     def ApplyToText(transform: Callable[[str],Tuple[str,int]]) -> int:
         changeCount = 0
@@ -47,7 +62,7 @@ def RenderDocumentationFiles(aboutDir: str,destDir:str = "",pathToPrototype:str 
     if html:
         htmlFiles = {}
         for fileName in fileContents:
-            html = markdown.markdown(fileContents[fileName],extensions = ["sane_lists","footnotes",NewTabRemoteExtension()])
+            html = markdown.markdown(fileContents[fileName],extensions = ["sane_lists","footnotes","toc",NewTabRemoteExtension()])
         
             html = re.sub(r"<!--HTML(.*?)-->",r"\1",html) # Remove comments around HTML code
             htmlFiles[Utils.ReplaceExtension(fileName,".html")] = html
@@ -65,22 +80,68 @@ def RenderDocumentationFiles(aboutDir: str,destDir:str = "",pathToPrototype:str 
 
         page = Html.PageDesc(Html.PageInfo(title,Utils.PosixJoin(destDir,fileName),titleInPage))
         page.AppendContent(fileText)
+        page.sourceFile = Utils.PosixJoin(sourceDir,Utils.ReplaceExtension(fileName,".md"))
         renderedPages.append(page)
 
     return renderedPages
 
+def PrintWordCount() -> None:
+    "Calculate the number of words in the text of the archive."
+
+    def CountMutipleTexts(texts: Iterable[str]) -> int:
+        words = 0
+        for text in texts:
+            words += WordCount(text)
+        return words
+    
+    wc = {}
+    wc["Excerpt"] = CountMutipleTexts(item["text"] for item in (itertools.chain.from_iterable(Filter.AllItems(x) for x in gDatabase["excerpts"])))
+    wc["Event description"] = CountMutipleTexts(e["description"] for e in gDatabase["event"].values())
+    wc["Documentation"] = gDocumentationWordCount
+    wc["Total"] = sum(wc.values())
+
+    for name in wc:
+        Alert.info(f"{name} word count: {wc[name]}")
 
 def AddArguments(parser) -> None:
     "Add command-line arguments used by this module"
     parser.add_argument('--documentationDir',type=str,default='documentation',help='Read and write documentation files here; Default: ./documenation')
+    parser.add_argument('--info',type=str,action="append",default=[],help="Specify infomation about this build. Format key:value")
     
 
+def ParseArguments() -> None:
+    class NameSpace:
+        pass
+    infoObject = NameSpace()
+    for item in gOptions.info:
+        split = item.split(":")
+        if len(split) > 1:
+            value = split[1]
+        else:
+            value = True
+        setattr(infoObject,split[0],value)
+    gOptions.info = infoObject
+    
+
+def Initialize() -> None:
+    pass
+
 gOptions = None
-gDatabase:dict = {} # These globals are overwritten by QSArchive.py, but we define them to keep PyLint happy
+gDatabase:dict[str] = {} # These globals are overwritten by QSArchive.py, but we define them to keep PyLint happy
+gDocumentationWordCount = 0
 
 def main() -> None:
+    global gDocumentationWordCount
+    gDocumentationWordCount = 0
+    modifiedFiles = []
     for directory in ['about','misc']:
         os.makedirs(Utils.PosixJoin(gOptions.documentationDir,directory),exist_ok=True)
         for page in RenderDocumentationFiles(directory,pathToPrototype=Utils.PosixJoin("../../",gOptions.prototypeDir),pathToBase="../../",html=False):
-            with open(page.info.file,'w',encoding='utf-8') as file:
-                print(str(page),file=file)
+            if not os.path.isfile(page.info.file) or Utils.DependenciesModified(page.info.file,[page.sourceFile]):
+                with open(page.info.file,'w',encoding='utf-8') as file:
+                    print(str(page),file=file)
+                modifiedFiles.append(page.info.file)
+    
+    if Alert.verbosity >= 2:
+        PrintWordCount()
+    Alert.info("Wrote",len(modifiedFiles),".md files:",modifiedFiles)
